Validate recorded-data learner updates on four-chip TPU with FSDP

+++ expo_ft/networks/temperature.py
@@ -9,6 +9,8 @@ class Temperature(nn.Module):
     def __call__(self) -> jnp.ndarray:
         log_temp = self.param(
             "log_temp",
-            init_fn=lambda key: jnp.full((), jnp.log(self.initial_temperature)),
+            # Explicit dtype keeps this parameter and Adam's moments strongly
+            # typed, avoiding JIT signature changes in the first two updates.
+            init_fn=lambda key: jnp.full((), jnp.log(self.initial_temperature), dtype=jnp.float32),
         )
         return jnp.exp(log_temp)
